data-engineer-challenge/data_processor/requirements.txt
paho-mqtt
pydantic[dotenv]
pydantic-settings
pyspark==3.5.4
python-dotenv
